cpychecker: add test coverage for Reporter.to_json
Split up check_refcounts(), doing the bulk of the work within a new
impl_check_refcounts(), which generates a Reporter instance holding
the unprocessed reports.

Use this to add test coverage for Reporter.to_json within
 tests/cpychecker/refcounts/json/basic/script.py
converting this test from a test of the JSON for a trace's end-state
into one for the reports found on the code.