cpychecker: avoid hardcoding the maximum number of transitions
Previously, the maximum number of transitions to analyze was hardcoded as 256,
which is good for simple examples, but makes the checker give up on complicated
functions.

With this commit this limit is now exposed as "maxtrans" all the way up the API,
so it can be set by individual test cases.

It is also now exposed as a --maxtrans option to gcc-with-cpychecker