show state changes when reporting an error
Add ExplodedGraph.get_shortest_path() to get the shortest path from the
initial (location, state) to a target (location, state), and use this within
the error() hook to determine the simplest way to reach the error state.

Use this when reporting an error to also report the nodes within the path at
which state changes happen, on the grounds that these are the most pertinent
aspects of the path.