refactor ExplodedGraph
Introduce gccutils/graph.py and rewrite ExplodedGraph in terms of it.

Introduce gccutils/dot.py and eliminate ExplodedGraphPrettyPrinter, given that
the latter had no meaningful state.

The result printer now has a path as a list of Edges, rather than as a list of
Nodes.

This should eventually make it possible to examine the actual transitions that
happen on a path.