provide human-readable descriptions of the path that led to a bug
When building the ExplodedGraph, use uniqueness of the the
  (srcexpnode, dstexpnode, pattern)
triple, rather than just the
  (srcexpnode, dstnode)
pair, and store the pattern (if any) within each ExplodedEdge, capturing the
pattern matches that lead to a particular path.

Add a description() method to the Pattern classes, giving a human-readable
string from an sm.solverContext, and use it when reporting errors to describe
the state transitions on the route to a bug occurring.

Add a context arg to the various gccutils.graph to_dot methods, and use this
to pass the sm.solver.Context to an ExplodedGraph when converting it to dot.

Fix ResultOfFnCall.matched_by() so that it checks for a match on the
left-hand-side.

Consistently pass strings for dot through to_html()

Fix some "sm.parser" references which should have been "sm.checker"