work towards interprocedural analysis
Rework the sm.solver.make_exploded_graph() to work on a Graph, rather than
on the function's CFG, passing in the StmtGraph of the function.

Extend the StmtGraph and Supergraph class families to support the needed
API, adding get_entry_nodes() to the graph class, get_stmt() and
get_gcc_loc() to the node classes, and true_value/false_value properties
to the edge classes.

Add beginnings of an interprocedural implementation of the same, though in
disabled form.