reimplement sm code to support interprocedural analysis
Reimplement the sm checker code as an interprocedural pass, working once on
the supergraph, rather than on each function's CFG in turn, so that we can
track e.g. where a ptr is malloced in one function, and accessed without a
check against NULL in another.

Reimplement the solver code to work on all possible states at once, rather
than on a single variable at a time, so that we can track pointer values
entering and leaving functions, handling interprocedural edges appropriately.
Add State and ShapeVar classes to solver to track pointer aliases.

Remove the "var" data from Context, and add it to a new Match class; change
Pattern.matched_by to Pattern.iter_matches

Don't treat static functions as possible entrypoints

Store errors as the graph is built, and play them back later, so that we can
emit them in source-order; add support for < <= > >= to gcc.Location so that
they can be sorted.

Add an AssignmentFromLiteral pattern and use it to match "q = 0" so that q can
be marked as NULL.