add leak detection
Add parser support for matching on "special patterns" such as:
  $leaked$
which are implemented as specific Pattern subclasses in sm.checker;
implement LeakedPattern as one of these.

Add _aliases to sm.solver.Shape to track all of the vars (if any) pointing
at a ShapeVar.  Introduce a ShapeChange class to capture the difference
between old and new Shapes and use it for all changes that involve a change
in aliasing.

Implement ShapeChange.iter_leaks() to supply all ShapeVars in the srcshape
that are no longer referenced in the dstshape.

Add Pattern.iter_expedge_matches, for patterns that match (potentially) on
exploded edges, rather than on nodes, and wire it up to be called by
ExplodedGraph.lazily_add_edge()

Add this clause to malloc_checker.sm:

  ptr.unknown, ptr.nonnull:
      $leaked$ => { error('leak of %s' % ptr)};

Wire it all up, so that the LeakedPattern is run for exploded edges that
have a ShapeChange in which pointer variables in those states leak, thus
leading to the error message.

Bulletproof error-handling for the case where the node has no gcc.Location,
using the end of the relevant function for this case (the ShapeChange
happens at the exit of the function, so it works for this case).

Also, add Options class to Context, so that we can disable caching of Error
instances (it's much easier to debug tracebacks with errors if we can emit
them immediately)