port solver to work on SSA representation
Running in LTO works on the SSA representation: I've been testing LTO analysis
by running the analysis before the "whole-program" IPA pass.

For sanity's sake we should minimize the differences between the LTO and
non-LTO modes of analysis.

Hence this commit ports the analysis to run before the "whole-program" IPA
pass, on the SSA representation, in preparation for LTO support.