add an example of custom state (tracking sizes within sizeof_allocation)
Generalize AssignmentFromLiteral to Assignment, to allow matching on
assignments, and fix the solver's general-purpose handling for assignments
so patterns can be matched after it is run.  Doing this required tweaking
Context.compare() to avoid it trying to match numeric literals in patterns
against gcc variables; it now only tries to match IDs within the patterns.

Add set_state(name, **kwargs) function to the Python API, allowing extra
key/value pairs to be added to a state (and disabling the check for
reachable states for checkers where this is used).

Expose the current state to the Python API, as "state".

Use the above to extend sizeof_allocation.sm to cover the case where the
cast to the destination type happens on a different line to that of the
allocation, which requires setting a custom state (with a "size" attribute)

Add a test case for the above:
   tests/sm/checkers/sizeof_allocation/stateful_malloc_too_small