autodetect whether the plugin needs to be built with C++
Recent builds of gcc 4.7 are built with C++, rather than C, and so the
internal symbols are name-mangled.

Try to determine whether or not the GCC we're building against was built with
C++; if so, supply "-x c++" during the build of the plugin.

Generate a new autogenerated-EXTRA_CFLAGS.txt containing additional flags
needed during compilation, so that this can hold "-x c++" if needed.