add integration test for the sm code
Create a "slow-tests" subdirectory for selftests that take much longer than
the rest of the test suite:

   python run-test-suite slow-tests

Create the first such test:
  slow-tests/sm/integration/lemon
containing a copy of lemon.c from wireshark-1.8.3.tar.bz2:tools/lemon/lemon.c
to give an example of a real non-trivial C file for the checker to analyze.

Currently it can't handle it.