start adding unit tests for the solver
Support passing a selftest callback to sm.main(), running it on the Context
and Solution after the solver has finished.

Add an initial example of such a selftest, verifying that a particular state
change is captured by the fixed point solver.

Add utility methods to sm.Context to simplify the selftest.