rework the fixed point solver to work on equivalence classes of expressions
Rewrite AbstractValue so that rather than tracking states for each
expression at a node, it tracks them for each *equivalence class* of
expression at a node.

Doing this ties together the states of the various temporaries and their
"real" expressions.  This also means that an AbstractValue is for a
particular node, since the equivalence classes only make sense w.r.t. a
given node.

Doing this means that (finally), we can prove that within
tests/sm/checkers/taint/example:test3 that "tmp.i" is in the x.ok state
when used in the index lookup: we know that it has been checked against
upper and lower bounds, despite those checks actually happening on
temporaries - those temporaries are within the same equivalence class as
"tmp.i".

Verify that "tmp.i" is uniquely in state "x.ok" for that test case, adding
various unit test support methods.

This is similar to 20fd591ae40e7e03b1f7e2935ec211a67e906d2a (which was for
the old solver), but is for the fixed point solver.  Note that even with
that commit, the old solver was only able to establish that tmp.i had the
same possible states as the temporaries, but because only one variable was
tracked at a time, each time "tmp.i" was looked up for a temporary it got
the start state, leading to a re-injection of "x.tainted" into the state set
each time, thus getting the wrong state set for that equivcls.

The new solver tracks combinations of states, in particular both "tmp" and
"tmp.i" and thus is able to narrow down the state of the equivalence class
to "x.ok" (the correct answer)

Note that the new solver is still not yet wired up to the error generator
(the old solver is used for that)