eliminate the noise when a candidate error is found to be impossible
Work towards JSON dumps of the results: rather than directly generate gcc
errors and notes for an Error, capture the description as a Report instance

Introduce a StderrReporter class to handle generating the faked:
  In function 'foo'
headers, and only start emitting an error is a full Report is available.

This eliminates the "In function 'foo'" noise for errors that are found to
be impossible.