reimplement error emission by walking the exploded graph
Previously, for each error we would construct an ErrorGraph, and run the
fact-finder over it, pruning impossible edges, then try to find a path from
an entrypoint to the node where the error occurs.  This implied a lot of
per-error work.

With this commit, we instead use the ExplodedGraph of states, and run the
fact-finder (and pruning) once on that.  We then use this per-error to find
a path (if any) from an entrypoint to the error site.

In order to generate meaningful error reports we also now backtrack along
the path to locate statechanges of interest.  This wasn't needed before
since the old implementation generated the ErrorGraph by backtracking
through the graph finding the interesting nodes as it went.