apply outcomes for a pattern match in sequence
Running "make demo-sm-lemon-timing" showed large numbers of duplicate
errors, where the checker would issue a report for every usage of a
variable in a bad state, rather than just the first such usage.

Update malloc_checker.sm so that after each error it transitions to another
state to avoid issuing multiple reports for the same underlying error.

Add selftests for this:
  tests/sm/checkers/malloc-checker/multiple-errors/deref-of-null
  tests/sm/checkers/malloc-checker/multiple-errors/unchecked_malloc

Previously the implementation considered multiple outcomes for a pattern
match as independent, running all of them effectively in parallel, whereas
we want them all to be run in sequence.

Update PossibleMatch and find_possible_matches() so that a PossibleMatch
holds the list of outcomes for a pattern rule that could apply at a given
edge, rather than building multiple PossibleMatch instances per rule, one
per outcome.

Generalize PythonEffect to just Effect, and replace Outcome.get_result()
with Outcome.get_effect_for_state(), eliminating FixedPointMatchContext.
Update both places that apply outcomes to handle a sequence of outcomes
for a pattern rule.

Doing so reduces the number of errors for "make demo-sm-lemon-timing"
from 94 to 34.