fixes to dynamic linking
Previously python.so was dynamically linked against "gcc-c-api/gcc-c-api.so", leaving a:
   NEEDED            Shared library: [gcc-c-api/gcc-c-api.so]
within its ELF metadata, which can only be resolved from a working copy.

Rename the library to libgcccapi.so, and fix things so that we link the
plugin against -lgcccapi, passing in -L so that it can find it when
python.so is linked.

This requires us to pass in LD_LIBRARY_PATH when invoking the plugin from
a working copy (or from elsewhere in the filesystem), but does allow the
plugin to be invoked from an arbitary location in the filesystem
(we may need to add an RPATH to python.so for installation).