revamp gcc-python-plugin.spec so that it uses builds
233920649d310a0210e37b2845b77b2c8a531046 moved the installation code from
the rpm specfile's %install clause to a "make install" target in the
Makefile.

Doing this broke the rpm build from the specfile.

This was because each of the builds were sharing the same builddir, which
meant that each "make install" rebuilt the .so file, but used the last set
of .o files that it found, leading to all of the installs using the .o files
from the final build, and thus all using the same Python runtime (the last
one to be built).

The fix is to use 4 different builddirs.

The ideal would be to share the srcdir and have 4 different builddirs.
Unfortunately the plugin doesn't yet cleanly support srcdir != builddir.
I tried fixing that, but the results got increasingly messy

So for now we simply make 4 separate copies of the source, one for each
build.

Also, comment-out the selftests for now, as they don't pass and this breaks
the build