Extract gcc version without running a test program
Change generation of print-gcc-version to preprocess, but not compile, a
dummy source file containing the macro GCC_VERSION, then scan the
resulting output to find the numeric value of the macro.  This lets the
compiler do the work of expanding the expression, but avoids running a
test program.  This is necessary for cross-compiler support, since
cross-compiled test outputs cannot be run locally.

Store the output in autogenerated-gcc-version, since it is not a
runnable program.