libcpychecker: Switch off verify_refcounting for gcc-7 and later
Refcounting verification fails in some cases for gcc-7 and later.

Workaround this by disabling refcounting verification entirely for gcc-7 and
later, to make sure that users don't run into python exceptions.

Emit warning in gcc-with-cpychecker when refcounting verification has been
disabled, to make sure that users are aware of this.