gcc-with-cpychecker: Add CC_FOR_CPYCHECKER (#136)
gcc-with-cpychecker does not use CC to find gcc, in order to avoid forkbombing
when setting CC=gcc-with-cpychecker, and instead uses a hardcoded 'gcc'.  This
means that when we want to use a specific gcc, we either needs to edit
gcc-with-cpychecker to adjust the filename, or use PATH settings to pick up
gcc (which may mean having to map somehow f.i. 'gcc-7' to 'gcc').

Add handling of CC_FOR_CPYCHECKER in gcc-with-cpychecker, to be able to easily
specify the gcc to be used.