Makefile: Propagate correct srcdir to gcc-c-api submake
If srcdir=/a/b is defined in the environment or on the make command line, then
srcdir is inherited with the same value in a submake.  However, the use of
srcdir in gcc-c-api/Makefile indicates that the expected value is
/a/b/gcc-c-api.

Fix this by propagating srcdir explicitly on the submake command line.