DRAFT removing the Wrapper metaclass for python 3.9
This code here is experimental.
I am using the docker image zbeekman/nightly-gcc-trunk-docker-image
latest 5114095361e6  with patches. gcc (GCC) 10.0.0
20191203 (experimental) latest 5114095361e6  I guess that
is actually older than I thought but whatever.
I am testing with pyenv 3.9-dev Python 3.9.9 .

The first thing I did was hook it up to the new python, then I had to
remove some deprecated code. Finally I worked around the issue of the
metaclass not loading. It does not even seem to be used so I am
removing it for now until I can learn more.

The goal here is to have a docker container that will contain the
latest version of the python plugin in a way that anyone can use it
easily with a docker pull.

Merry Christmas and Happy New Year 2020

mike